services/controller-api/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "penguinwhisk-controller"
version = "0.1.0.0"
description = "PenguinWhisk Controller API - OpenWhisk ARM deployment controller"
requires-python = ">=3.13"
authors = [
    {name = "Penguin Tech Inc", email = "dev@penguintech.io"}
]
license = {text = "Limited AGPL3"}
dependencies = [
    "flask>=3.0.0",
    "flask-security-too>=5.4.0",
    "pydal>=20240101",
    "redis>=5.0.0",
    "minio>=7.2.0",
    "gunicorn>=21.0.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
    "python-dotenv>=1.0.0",
    "marshmallow>=3.20.0",
    "psycopg2-binary>=2.9.9",
    "pymysql>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "flake8>=6.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "bandit>=1.7.5",
    "types-redis>=4.6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "tests*"]

[tool.black]
line-length = 100
target-version = ["py313"]
include = '\.pyi?$'
extend-exclude = '''
/(
  \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = [".tox/*", "venv/*"]
python_version = "313"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
plugins = ["flask_sqlalchemy_stubs"]

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-v --strict-markers --tb=short --cov=app --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "e2e: end-to-end tests",
]

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101"]
